requirements.txt
streamlit
google-genai
faiss-cpu
langchain
langchain-community
langchain-huggingface
sentence-transformers